requirements.txt
baidu-aip
pyaudio